# use fast csv parser for more input types