--- a/polars/polars-io/src/mmap.rs
+++ b/polars/polars-io/src/mmap.rs
@@ -1,0 +1,39 @@
+use std::fs::File;
+use std::io::{Cursor, Read, Seek};
+
+/// Trait used to get a hold to file handler or to the underlying bytes
+/// without performing a Read.
+pub trait MmapBytesReader: Read + Seek + Send + Sync {
+    fn to_file(&self) -> Option<&File> {
+        None
+    }
+
+    fn to_bytes(&self) -> Option<&[u8]> {
+        None
+    }
+}
+
+impl MmapBytesReader for File {
+    fn to_file(&self) -> Option<&File> {
+        Some(self)
+    }
+}
+
+impl<T> MmapBytesReader for Cursor<T>
+where
+    T: AsRef<[u8]> + Send + Sync,
+{
+    fn to_bytes(&self) -> Option<&[u8]> {
+        Some(self.get_ref().as_ref())
+    }
+}
+
+impl<T: MmapBytesReader + ?Sized> MmapBytesReader for Box<T> {
+    fn to_file(&self) -> Option<&File> {
+        T::to_file(self)
+    }
+
+    fn to_bytes(&self) -> Option<&[u8]> {
+        T::to_bytes(self)
+    }
+}