--- a/polars/polars-lazy/src/physical_plan/executors/mod.rs
+++ b/polars/polars-lazy/src/physical_plan/executors/mod.rs
@@ -17,7 +17,7 @@ use crate::logical_plan::FETCH_ROWS;
 use itertools::Itertools;
 use polars_core::POOL;
 use rayon::prelude::*;
-use std::io::{Read, Seek};
+use std::io::Seek;
 use std::path::PathBuf;
 
 const POLARS_VERBOSE: &str = "POLARS_VERBOSE";