@ghuls @olivier-lacroix 

This will probably also improve performance of `ffspec` related code. All csv-parsing will use multiple threads and in memory buffers as `bytes` of `io.BytesIO` will have no redundant copy. 

Next up lazy csv parsing.