@olivier-lacroix 2021-07-11
@ritchie46 , it seems there is no overhead at all for local files when using fsspec. Is this real, or is the rust code detecting somehow this is a local file and bypasses python entirely ?

```
with fsspec.open(local_file) as f:
    pl.read_csv(f, use_pyarrow=False)
# 2.06 s ± 16.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

pl.read_csv(local_file, use_pyarrow=False)
# 2.06 s ± 23 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

Opening the same file stored in a GCS bucket is still much slower, but this PR does increase performance:

```
# before this PR
pl.read_csv(remote_file, use_pyarrow=False)
# 8.8 s ± 283 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

# after this PR
pl.read_csv(remote_file, use_pyarrow=False)
# 7.34 s ± 337 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

For comparison, equivalent operation in pandas takes 17s when the file is stored in GCS, and 10s when the file is local.

Tests done with a 2 vCPU machine.