@ritchie46 2021-07-11
> @ritchie46 , it seems there is no overhead at all for local files when using fsspec. Is this real, or is the rust code detecting somehow this is a local file and bypasses python entirely ?

This is as expected. If I can still get a hold of the original filename, we skip python entirely and open the file via `mmap`.

Another fast path is when we have an in memory buffer in python. If we can turn this into a bytes slice `&[u8]`, then we also skip python completely and work on the memory address of the buffer. This is the case for `io.BytesIO, io.TextIO, bytes`, and probably a few more types.

If we do have to deal with a python object, that's not an in memory buffer, we have to call `read` which is interaction with python. In that case we first read all into a Rust memory buffer and then continue multi-threaded on that buffer.

The performance improvement on the GCS bucket is due to multi-threading. We still have to download into a buffer, so that is not sped up.

But given that pandas takes 17s for a remote file, I am quite happy with this approach. :)

Note: there will be a future speed up for parsing string data. See #955