@olivier-lacroix 2021-07-12
@ritchie46, indeed. Do you see a way to force the python path for local files and therefore quantify the penalty of going through fsspec / python ?

Yes, the result on GCS is really good! And the penalty of reading from gcs is even lower in polars than in pandas: ~5s versus ~7s, which is quite surprising.

Also, I don't think gcsfs (that is used behind the scene by fsspec) is as performant as it could be: reading a 700Mo CSV should most likely be achieved in less than a second based on expected bandwidth.