@ritchie46 2021-07-12
> Already, using the official google python client is faster than gcsfs (saving 1.8s or 30% of data transfer time):

```python
from google.cloud import storage

client = storage.Client()
bucket = client.get_bucket('bucket')
blob = bucket.get_blob('filename')

# dowload as bytes
data = blob.download_as_bytes()
pl.read_csv(data, use_pyarrow=False)
# 5.56 s ± 199 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```

This is actually how I'd like to see it. Then both tools can optimize where they are good at. GCS optimizes their transfer, we optimize the parsing. And its explicit. :slightly_smiling_face: