@olivier-lacroix 2021-07-14
Fair enough. But that prevents streaming: ideally processing should start as soon as first bytes are received.