@ritchie46 2021-07-14
> Fair enough. But that prevents streaming: ideally processing should start as soon as first bytes are received.

Yes, but then I'd need to write a sequential read based algorithm for the csv-parser as well. I don't want the extra code maintenance for potential diminishing returns. I believe pandas csv-parser is read based and thus streaming.