@ghuls 2021-07-14
@olivier-lacroix Does reading gzipped files work for you (fsspec is installed)?

This results in reading compressed data:
```python
pl.read_csv('test.tsv.gz', use_pyarrow=False)
```

@ritchie46 Can it be that you are reading the compressed file directly even when `fsspec` deternined that the file was compressed?