@olivier-lacroix 2021-07-15
@ghuls , you are right, it does not work anymore.  This may be because of [this ](https://github.com/pola-rs/polars/pull/943/files#diff-4766be397c9d03cfe00302fbcfcc46f30222df2b039a66534a86a0a3c568441cR248-R252) branch ?