@ritchie46 2021-07-15
> @ritchie46 Can you add a check there to see if there is a compression attribute set (py_f.getattr("compression") and only when it is None, try to open the file in Rust directly (until reading compressed files in Rust itself is supported).

Yes, will do!