@ritchie46 2021-07-15
Could someone add a test, so that this regression will not occur in the future?