@ghuls 2021-07-15
It still seems broken for me after https://github.com/pola-rs/polars/commit/53da7755a9af8dc29461f41f477d86a7b59d00d7